test(catalog): run handoff lifecycle in real Chromium
