+++ apps/agor-ui/src/components/Marketplace/MCPCatalogModal.test.tsx
@@ -9,7 +9,7 @@ import {
   waitFor,
   within,
 } from '@testing-library/react';
-import { afterEach, describe, expect, it, vi } from 'vitest';
+import { afterEach, describe, expect, it } from 'vitest';
 import {
   CatalogHarness,
   catalogUser,
@@ -130,39 +130,8 @@ describe('app-owned MCP Catalog', () => {
 });
 
 describe('onboarding Catalog handoff', () => {
-  it('opens the reviewed PAT drawer for the new branch, never connects eagerly, and forgets it on close', async () => {
-    const api = makeCatalogClient([githubHandoffEntry]);
-    const consumed = vi.fn();
-    // This jsdom case tests a continuous close/reopen lifecycle, not animation.
-    // Browser fixtures retain motion; keep the normal timeout and real components.
-    render(
-      <CatalogHarness
-        motion={false}
-        client={api.client}
-        handoff={{ entryName: githubHandoffEntry.name, branchId: 'branch-1' }}
-        onHandoffConsumed={consumed}
-      />
-    );
-    const drawer = await screen.findByRole('dialog', { name: /GitHub/ });
-    await within(drawer).findByText('Catalog QA');
-    expect(
-      within(drawer).getByPlaceholderText('Paste your GitHub bearer access token')
-    ).toHaveValue('');
-    expect(api.connect).not.toHaveBeenCalled();
-    expect(consumed).toHaveBeenCalledTimes(1);
-    fireEvent.click(within(drawer).getByRole('button', { name: 'Close' }));
-    // Assert actual teardown, not a temporarily inaccessible animated node.
-    await waitFor(() => expect(drawer).not.toBeInTheDocument());
-    expect(screen.queryByRole('dialog', { name: /GitHub/ })).not.toBeInTheDocument();
-    expect(api.connect).not.toHaveBeenCalled();
-    const modal = screen.getByRole('dialog', { name: 'MCP Catalog' });
-    fireEvent.click(within(modal).getByRole('button', { name: 'Close' }));
-    await waitFor(() => expect(modal).not.toBeInTheDocument());
-    expect(screen.queryByRole('dialog')).not.toBeInTheDocument();
-    fireEvent.click(screen.getByRole('button', { name: 'Open MCP Catalog' }));
-    await screen.findByRole('dialog', { name: 'MCP Catalog' });
-    expect(screen.queryByRole('dialog', { name: /GitHub/ })).not.toBeInTheDocument();
-  });
+  // The continuous handoff close/reopen lifecycle runs with real motion in
+  // MCPCatalogModal.browser.test.tsx; keep state/authority cases here.
 
   it('does not silently use another branch if the onboarding branch is not visible', async () => {
     const api = makeCatalogClient([githubHandoffEntry]);
